Extracts createProject to own file for testing
